requirements.txt
Django==1.11.21
psycopg2==2.7.3.1
python-decouple==3.1
pytz==2017.2
